# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transport-options": {
      "get": {
        "operationId": "TransportController_options",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/trade-orders": {
      "get": {
        "operationId": "TransportController_orders",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "businessEntityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grainId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按主体查询交易订单明细及剩余可运输数量",
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/trade-orders/import": {
      "post": {
        "operationId": "TransportController_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "平台幂等导入交易订单，不覆盖已有来源事实",
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-demands": {
      "get": {
        "operationId": "TransportController_demands",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "businessEntityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grainId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      },
      "post": {
        "operationId": "TransportController_createDemand",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-demands/{id}": {
      "get": {
        "operationId": "TransportController_demand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      },
      "put": {
        "operationId": "TransportController_updateDemand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-demands/{id}/publish": {
      "post": {
        "operationId": "TransportController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishDemandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-demands/{id}/{action}": {
      "post": {
        "operationId": "TransportController_demandAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-demands/{id}/history": {
      "get": {
        "operationId": "TransportController_demandHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-supplies": {
      "get": {
        "operationId": "TransportController_supplies",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "businessEntityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grainId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      },
      "post": {
        "operationId": "TransportController_createSupply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-supplies/{id}": {
      "get": {
        "operationId": "TransportController_supply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      },
      "put": {
        "operationId": "TransportController_updateSupply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-supplies/{id}/publish": {
      "post": {
        "operationId": "TransportController_publishSupply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-supplies/{id}/{action}": {
      "post": {
        "operationId": "TransportController_supplyAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
    "/api/transport-supplies/{id}/history": {
      "get": {
        "operationId": "TransportController_supplyHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "北粮南运 · 智能多式联运 API",
    "description": "Phase 9：需求、方案、匹配、模拟签章、联运执行、预测预警、智能辅助、账单版本、双方结算确认和授权数据服务。",
    "version": "0.9.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "lineNo": {
            "type": "string"
          },
          "grainId": {
            "type": "string"
          },
          "cargoName": {
            "type": "string"
          },
          "specification": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "lineNo",
          "grainId",
          "cargoName",
          "specification",
          "quantity"
        ]
      },
      "ImportOrderDto": {
        "type": "object",
        "properties": {
          "businessNo": {
            "type": "string"
          },
          "businessEntityId": {
            "type": "string"
          },
          "recipient": {
            "type": "string"
          },
          "shipperContact": {
            "type": "string"
          },
          "shipperPhone": {
            "type": "string"
          },
          "recipientContact": {
            "type": "string"
          },
          "recipientPhone": {
            "type": "string"
          },
          "sourceSystem": {
            "type": "string"
          },
          "sourceRecordId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "businessNo",
          "businessEntityId",
          "recipient",
          "shipperContact",
          "shipperPhone",
          "recipientContact",
          "recipientPhone",
          "sourceSystem",
          "sourceRecordId",
          "items"
        ]
      },
      "DemandDto": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "originCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "originAddress": {
            "type": "string"
          },
          "destinationCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "destinationAddress": {
            "type": "string"
          },
          "departureAt": {
            "type": "string"
          },
          "arrivalAt": {
            "type": "string"
          },
          "modeIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowMultimodal": {
            "type": "boolean"
          },
          "allowTransfer": {
            "type": "boolean"
          },
          "maxTransfers": {
            "type": "number"
          },
          "loadingType": {
            "type": "string"
          },
          "preference": {
            "type": "string"
          },
          "budget": {
            "type": "number"
          },
          "contact": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "orderItemId",
          "name",
          "quantity",
          "originCodes",
          "originAddress",
          "destinationCodes",
          "destinationAddress",
          "departureAt",
          "modeIds",
          "allowMultimodal",
          "allowTransfer",
          "maxTransfers",
          "loadingType",
          "preference",
          "contact",
          "phone"
        ]
      },
      "PublishDemandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number"
          },
          "allowPartial": {
            "type": "boolean"
          },
          "mode": {
            "type": "string"
          },
          "targetCarrierId": {
            "type": "string"
          },
          "deadline": {
            "type": "string"
          },
          "quoteType": {
            "type": "string"
          },
          "budgetPublic": {
            "type": "boolean"
          },
          "contactPublic": {
            "type": "boolean"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "version",
          "mode",
          "deadline",
          "quoteType",
          "budgetPublic",
          "contactPublic"
        ]
      },
      "VersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ]
      },
      "SupplyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "modeId": {
            "type": "string"
          },
          "originCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "originAddress": {
            "type": "string"
          },
          "destinationCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "destinationAddress": {
            "type": "string"
          },
          "viaNodes": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "minQuantity": {
            "type": "number"
          },
          "maxQuantity": {
            "type": "number"
          },
          "resourceType": {
            "type": "string"
          },
          "resourceDescription": {
            "type": "string"
          },
          "grainIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "serviceStart": {
            "type": "string"
          },
          "serviceEnd": {
            "type": "string"
          },
          "durationHours": {
            "type": "number"
          },
          "referencePrice": {
            "type": "number"
          },
          "priceUnit": {
            "type": "string"
          },
          "capabilities": {
            "type": "string"
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "modeId",
          "originCodes",
          "originAddress",
          "destinationCodes",
          "destinationAddress",
          "capacity",
          "resourceType",
          "resourceDescription",
          "grainIds",
          "serviceStart",
          "serviceEnd",
          "validFrom",
          "validUntil",
          "contact",
          "phone"
        ]
      },
...
    }
  }
}